% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{bigobject-quoting}
\alias{bigobject-quoting}
\alias{dbQuoteIdentifier,BigObjectConnection,character-method}
\alias{dbQuoteIdentifier,BigObjectConnection,SQL-method}
\alias{dbQuoteString,BigObjectConnection,character-method}
\alias{dbQuoteString,BigObjectConnection,SQL-method}
\title{Quote BigObject strings and identifiers.}
\usage{
\S4method{dbQuoteIdentifier}{BigObjectConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{BigObjectConnection,SQL}(conn, x, ...)

\S4method{dbQuoteString}{BigObjectConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{BigObjectConnection,SQL}(conn, x, ...)
}
\description{
In BigObject, identifiers are enclosed in backticks, e.g. \code{`x`}.
}
\examples{
con <- dbConnect(RBigObject::BigObject())
dbQuoteIdentifier(con, c("a b", "a`b"))
dbQuoteString(con, c("a b", "a'b"))
dbDisconnect(con)
}
\keyword{internal}
