% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{bigobject-tables}
\alias{bigobject-tables}
\alias{dbReadTable,BigObjectConnection,character-method}
\alias{dbWriteTable,BigObjectConnection,character,data.frame-method}
\alias{dbListTables,BigObjectConnection-method}
\alias{dbExistsTable,BigObjectConnection,character-method}
\alias{dbRemoveTable,BigObjectConnection,character-method}
\title{Read and write BigObject tables.}
\usage{
\S4method{dbReadTable}{BigObjectConnection,character}(conn, name, ...,
  row.names = FALSE, check.names = TRUE)

\S4method{dbWriteTable}{BigObjectConnection,character,data.frame}(conn, name,
  value, field.types = NULL, row.names = NA, overwrite = FALSE,
  append = FALSE, ..., allow.keywords = FALSE, temporary = FALSE,
  default = TRUE)

\S4method{dbListTables}{BigObjectConnection}(conn, ...)

\S4method{dbExistsTable}{BigObjectConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{BigObjectConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{a \code{\linkS4class{BigObjectConnection}} object, produced by
\code{\link[DBI]{dbConnect}}}

\item{name}{a character string specifying a table name.}

\item{...}{Unused, needed for compatiblity with generic.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{check.names}{If \code{TRUE}, the default, column names will be
converted to valid R identifiers.}

\item{value}{A data frame.}

\item{field.types}{Optional, overrides default choices of field types,
derived from the classes of the columns in the data frame.}

\item{overwrite}{a logical specifying whether to overwrite an existing table
or not. Its default is \code{FALSE}. (See the BUGS section below)}

\item{append}{a logical specifying whether to append to an existing table
in the DBMS.  If appending, then the table (or temporary table)
must exist, otherwise an error is reported. Its default is \code{FALSE}.}

\item{allow.keywords}{DEPRECATED.}

\item{temporary}{If \code{TRUE}, creates a temporary table that expires
when the connection is closed.}

\item{default}{If \code{TRUE}, creates a default table in which rows can be deleted.}
}
\value{
A data.frame in the case of \code{dbReadTable}; otherwise a logical
indicating whether the operation was successful.
}
\description{
These functions mimic their R counterpart \code{get}, \code{assign},
\code{exists}, \code{remove}, and \code{ls}.
}
\note{
Note that the data.frame returned by \code{dbReadTable} only has
primitive data, e.g., it does not coerce character data to factors.
}
\examples{
con <- dbConnect(RBigObject::BigObject(), dbname = "test")

# By default, row names are written in a column to row_names, and
# automatically read back into the row.names()
dbWriteTable(con, "mtcars", mtcars[1:5, ], temporary = TRUE)
dbReadTable(con, "mtcars")
dbReadTable(con, "mtcars", row.names = NULL)
}
