% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigobject_sql.R
\name{bigobject_sql}
\alias{bigobject_sql}
\title{Submit SQL Query to BigObject}
\usage{
bigobject_sql(stmt, ip = getOption("BIGOBJECT_IP", "127.0.0.1"),
  port = getOption("BIGOBJECT_PORT", "9090"))
}
\arguments{
\item{stmt}{character. A series of SQL statement.
Each element is a complete SQL statement.
These query will be executed from the first element to the last element.}

\item{ip}{string. The ip address or domain name to the BigObject instance.}

\item{port}{string. The port number.}
}
\value{
A list with the following slots:
\itemize{
 \item{"Content"} The returned content.
 \item{"Status"}
 \item{"Err"}
}
}
\description{
Submit SQL Query to BigObject
}
\references{
\url{http://docs.bigobject.io/API/index.html}
}

