% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{bigobject_verbose}
\alias{bigobject_verbose}
\title{Configure verbose of RBigObject}
\usage{
bigobject_verbose(verbose)
}
\arguments{
\item{verbose}{logical. Whether to print verbose message.}
}
\description{
Configure verbose of RBigObject
}

