% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigObjectConnection.R
\docType{class}
\name{BigObjectConnection-class}
\alias{BigObjectConnection-class}
\alias{dbDisconnect,BigObjectConnection-method}
\alias{dbGetInfo,BigObjectConnection-method}
\alias{show,BigObjectConnection-method}
\alias{dbIsValid,BigObjectConnection-method}
\title{Class BigObjectConnection.}
\usage{
\S4method{dbDisconnect}{BigObjectConnection}(conn, ...)

\S4method{dbGetInfo}{BigObjectConnection}(dbObj, ...)

\S4method{show}{BigObjectConnection}(object)

\S4method{dbIsValid}{BigObjectConnection}(dbObj, ...)
}
\description{
\code{BigObjectConnection.} objects are usually created by
[DBI::dbConnect()]
}
\keyword{internal}
