% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigObjectDriver.R
\docType{class}
\name{BigObjectDriver-class}
\alias{BigObjectDriver-class}
\alias{dbUnloadDriver,BigObjectDriver-method}
\alias{dbIsValid,BigObjectDriver-method}
\alias{dbGetInfo,BigObjectDriver-method}
\title{Class BigObjectDriver with constructor BigObject.}
\usage{
\S4method{dbUnloadDriver}{BigObjectDriver}(drv, ...)

\S4method{dbIsValid}{BigObjectDriver}(dbObj, ...)

\S4method{dbGetInfo}{BigObjectDriver}(dbObj, what = "", ...)
}
\description{
An BigObject driver implementing the R database (DBI) API.
This class should always be initialized with the \code{BigObject()} function.
It returns a singleton that allows you to connect to BigObject.
}
\keyword{internal}
