% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigObjectResult.R
\docType{class}
\name{BigObjectResult-class}
\alias{BigObjectResult-class}
\alias{dbIsValid,BigObjectResult-method}
\title{Class BigObjectResult}
\usage{
\S4method{dbIsValid}{BigObjectResult}(dbObj, ...)
}
\description{
BigObject's query results class.  This classes encapsulates the result of an SQL
statement (either \code{select} or not).
}
\keyword{internal}
