% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{methods}
\name{dbConnect,BigObjectDriver-method}
\alias{dbConnect,BigObjectDriver-method}
\alias{BigObject}
\title{Connect/disconnect to a BigObject DBMS}
\usage{
\S4method{dbConnect}{BigObjectDriver}(drv, ...)

BigObject()
}
\arguments{
\item{drv}{an object of class \code{BigObjectDriver}, or the character string
"BigObject" or an \code{BigObjectConnection}.}

\item{...}{Unused, needed for compatibility with generic.}

\item{username, password}{Username and password. If username omitted,
defaults to the current user. If password is omitted, only users
without a password can log in.}

\item{dbname}{string with the database name or NULL. If not NULL, the
connection sets the default database to this value.}

\item{host}{string identifying the host machine running the BigObject server or
NULL. If NULL or the string \code{"localhost"}, a connection to the local
host is assumed.}

\item{port}{(optional) integer of the TCP/IP default port.}

\item{max.con}{DEPRECATED}

\item{fetch.default.rec}{DEPRECATED}
}
\description{
These methods are straight-forward implementations of the corresponding
generic functions.
}
\examples{
\dontrun{
# Connect to a BigObject database running locally
con <- dbConnect(RBigObject::BigObject(), dbname = "bigobject")
# Connect to a remote database with username and password
con <- dbConnect(RBigObject::BigObject(), host = "mydb.mycompany.com",port = 3306)
# But instead of supplying the username and password in code, it's usually

# Always cleanup by disconnecting the database
dbDisconnect(con)
}

# connect to a database and load some data
con <- dbConnect(RBigObject::BigObject(), dbname = "bigobject")
dbWriteTable(con, "USArrests", datasets::USArrests, overwrite = TRUE)

# query
rs <- dbSendQuery(con, "SELECT * FROM USArrests")
d1 <- dbFetch(rs, n = 10)      # extract data in chunks of 10 rows
dbHasCompleted(rs)
d2 <- dbFetch(rs, n = -1)      # extract all remaining data
dbHasCompleted(rs)
dbClearResult(rs)
dbListTables(con)

# clean up
dbRemoveTable(con, "USArrests")
dbDisconnect(con)
}
