% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{dbDataType,BigObjectConnection-method}
\alias{dbDataType,BigObjectConnection-method}
\alias{dbDataType,BigObjectDriver-method}
\title{Determine the SQL Data Type of an S object}
\usage{
\S4method{dbDataType}{BigObjectConnection}(dbObj, obj, ...)

\S4method{dbDataType}{BigObjectDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{A \code{BigObjectDriver} or \code{BigObjectConnection}.}

\item{obj}{R/S-Plus object whose SQL type we want to determine.}

\item{\dots}{any other parameters that individual methods may need.}
}
\description{
This method is a straight-forward implementation of the corresponding
generic function.
}
\examples{
dbDataType(RBigObject::BigObject(), "a")
dbDataType(RBigObject::BigObject(), 1:3)
dbDataType(RBigObject::BigObject(), 2.5)
}
