% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{methods}
\name{fetch,BigObjectResult,numeric-method}
\alias{fetch,BigObjectResult,numeric-method}
\title{Fetch rows (deprecated)}
\usage{
\S4method{fetch}{BigObjectResult,numeric}(res, n = -1, ...)
}
\description{
Please use \code{dbFetch} instead
}
\keyword{internal}
