% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBigObject.R
\docType{package}
\name{RBigObject-package}
\alias{RBigObject}
\alias{RBigObject-package}
\title{RBigObject is a database interface and 'BigObject' Driver}
\description{
RBigObject is a database interface and 'BigObject' Driver
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigobject-inc/RBigObject}
  \item \url{http://www.bigobject.io}
  \item Report bugs at \url{https://github.com/bigobject-inc/RBigObject/issues}
}

}
\author{
\strong{Maintainer}: Cheng-Ching Huang \email{cchuang@bigobject.io}

Authors:
\itemize{
  \item RMySQL authors
  \item RMariaDB authors
}

}
