% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{methods}
\name{result-meta}
\alias{result-meta}
\alias{dbColumnInfo,BigObjectResult-method}
\alias{dbListFields,BigObjectConnection,character-method}
\alias{dbGetRowsAffected,BigObjectResult-method}
\alias{dbGetRowCount,BigObjectResult-method}
\alias{dbHasCompleted,BigObjectResult-method}
\title{Database interface meta-data.}
\usage{
\S4method{dbColumnInfo}{BigObjectResult}(res, ...)

\S4method{dbListFields}{BigObjectConnection,character}(conn, name, ...)

\S4method{dbGetRowsAffected}{BigObjectResult}(res, ...)

\S4method{dbGetRowCount}{BigObjectResult}(res, ...)

\S4method{dbHasCompleted}{BigObjectResult}(res, ...)
}
\arguments{
\item{res}{An object of class \code{\linkS4class{BigObjectResult}}}

\item{...}{Ignored. Needed for compatibility with generic}
}
\description{
See documentation of generics for more details.
}
\examples{
con <- dbConnect(RBigObject::BigObject(), dbname = "bigobject")
dbWriteTable(con, "t1", datasets::USArrests, overwrite = TRUE)

rs <- dbSendQuery(con, "SELECT * FROM t1 WHERE UrbanPop >= 80")
rs

dbGetStatement(rs)
dbHasCompleted(rs)
dbColumnInfo(rs)

dbFetch(rs)
rs

dbClearResult(rs)
dbRemoveTable(con, "t1")
dbDisconnect(con)
}
