% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigobject_sql.R
\name{bigobject_import}
\alias{bigobject_import}
\title{Import data.frame to BigObject}
\usage{
bigobject_import(df, name, action = c("create", "append", "overwrite"),
  ip = getOption("BIGOBJECT_IP", "127.0.0.1"),
  port = getOption("BIGOBJECT_PORT", "9090"),
  verbose = getOption("BIGOBJECT_VERBOSE", TRUE))
}
\arguments{
\item{df}{data.frame. The data which will be imported to BigObject.}

\item{name}{string. The table name of the imported}

\item{action}{string. Specify how to update object in BigObject. Only the first element is used. Please see details.}

\item{ip}{string. The ip address or domain name to the BigObject instance.}

\item{port}{string. The port number.}

\item{verbose}{logical value. Whether to print verbose message.}
}
\description{
Import data.frame to BigObject
}
\details{
The \code{action} parameter indicates the behavior of BigObject:
\itemize{
 \item {"create":} Creating a new table.
 \item {"append":} Appending data to an existed table.
 \item {"overwrite":} Overwrite an existed table with the same schema.
}
}
\references{
url{http://docs.bigobject.io/API/Data_Import_Service.html}
}

